refactor: put classes and related functions in dedicated files
